:sparkles: goal 스키마 구현
