+++ prisma/schema.prisma
@@ -40,6 +40,7 @@ model Schedule {
   UserId      Int
   createdAt   DateTime @default(now())
   updatedAt   DateTime @updatedAt
+
 }
 
 model Token {
@@ -113,6 +114,14 @@ model Notice {
   authorId  Int 
   study     Study    @relation(fields: [studyId], references: [id], onDelete: Cascade)
   studyId   Int
+  createdAt DateTime @default(now())
+  updatedAt DateTime @updatedAt
+}
+
+model Goal {
+  id        Int      @id @default(autoincrement())
+  time      Int
+  day       Int
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 }
