+++ src/components/MyPage/StudyChart.tsx
